Refactor out Photon presenter to make it default, separate geocode and component extensions
