+++ SpatiaBlazor.Demo/SpatiaBlazor.Demo/appsettings.Development.json
@@ -5,13 +5,15 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "SpatiaBlazor": {
-    "Geocode": {
-      "GooglePlaces": {
-        "ApiUrl": "https://places.googleapis.com/v1/places:autocomplete"     
-      },
-      "Photon": {
-        "ApiUrl": "https://photon.komoot.io/"
+  "Test": {
+    "SpatiaBlazor": {
+      "Geocode": {
+        "GooglePlaces": {
+          "ApiUrl": "https://places.googleapis.com/v1/places:autocomplete"
+        },
+        "Photon": {
+          "ApiUrl": "https://photon.komoot.io/"
+        }
       }
     }
   }
