+++ SpatiaBlazor/Components/Address/Suggestions/AddressSuggestionsExtensions.cs
@@ -1,0 +1,14 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace SpatiaBlazor.Components.Address.Suggestions;
+
+public static class AddressSuggestionsExtensions
+{
+    public static IServiceCollection AddAddressSuggestions(
+        this IServiceCollection services,
+        ServiceLifetime serviceLifetime = ServiceLifetime.Transient)
+    {
+        services.Add(new ServiceDescriptor(typeof(ISuggestionsPresenter), typeof(DefaultSuggestionsPresenter), serviceLifetime));
+        return services;
+    }
+}
