--- SpatiaBlazor/Components/Address/Suggestions/Photon/PhotonSuggestionsExtensions.cs
@@ -1,16 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-using SpatiaBlazor.Geocode.Photon;
-
-namespace SpatiaBlazor.Components.Address.Suggestions.Photon;
-
-public static class PhotonSuggestionsExtensions
-{
-    public static IServiceCollection AddPhotonAddressSuggestionsComponent(
-        this IServiceCollection services,
-        ServiceLifetime serviceLifetime = ServiceLifetime.Transient)
-    {
-        services.AddPhotonGeocodeClient(serviceLifetime);
-        services.Add(new ServiceDescriptor(typeof(ISuggestionsPresenter), typeof(PhotonSuggestionsPresenter), serviceLifetime));
-        return services;
-    }
-}
